+++ tests/conftest.py
@@ -1,28 +1,22 @@
 import os
-import json
-import base64
 import threading
 from pathlib import Path
-from typing import TYPE_CHECKING, Set, Dict, List, Union, TypeVar, Generator
+from typing import TYPE_CHECKING, Set, Generator
 
 import pytest
-from werkzeug import Request, Response
 from nonebug import NONEBOT_INIT_KWARGS
-from werkzeug.datastructures import MultiDict
 from werkzeug.serving import BaseWSGIServer, make_server
 
 import nonebot
 from nonebot.drivers import URL
+
+from .fake_server import request_handler
 
 os.environ["CONFIG_FROM_ENV"] = '{"test": "test"}'
 os.environ["CONFIG_OVERRIDE"] = "new"
 
 if TYPE_CHECKING:
     from nonebot.plugin import Plugin
-
-
-K = TypeVar("K")
-V = TypeVar("V")
 
 
 def pytest_configure(config: pytest.Config) -> None:
@@ -39,66 +33,6 @@ def load_plugin(nonebug_init: None) -> Set["Plugin"]:
 def load_builtin_plugin(nonebug_init: None) -> Set["Plugin"]:
     # preload builtin plugins
     return nonebot.load_builtin_plugins("echo", "single_session")
-
-
-def json_safe(string, content_type="application/octet-stream") -> str:
-    try:
-        string = string.decode("utf-8")
-        json.dumps(string)
-        return string
-    except (ValueError, TypeError):
-        return b"".join(
-            [
-                b"data:",
-                content_type.encode("utf-8"),
-                b";base64,",
-                base64.b64encode(string),
-            ]
-        ).decode("utf-8")
-
-
-def flattern(d: MultiDict[K, V]) -> Dict[K, Union[V, List[V]]]:
-    return {k: v[0] if len(v) == 1 else v for k, v in d.to_dict(flat=False).items()}
-
-
-@Request.application
-def request_handler(request: Request) -> Response:
-    try:
-        _json = json.loads(request.data.decode("utf-8"))
-    except (ValueError, TypeError):
-        _json = None
-
-    return Response(
-        json.dumps(
-            {
-                "url": request.url,
-                "method": request.method,
-                "origin": request.headers.get("X-Forwarded-For", request.remote_addr),
-                "headers": flattern(
-                    MultiDict((k, v) for k, v in request.headers.items())
-                ),
-                "args": flattern(request.args),
-                "form": flattern(request.form),
-                "data": json_safe(request.data),
-                "json": _json,
-                "files": flattern(
-                    MultiDict(
-                        (
-                            k,
-                            json_safe(
-                                v.read(),
-                                request.files[k].content_type
-                                or "application/octet-stream",
-                            ),
-                        )
-                        for k, v in request.files.items()
-                    )
-                ),
-            }
-        ),
-        status=200,
-        content_type="application/json",
-    )
 
 
 @pytest.fixture(scope="session", autouse=True)
